requirements.txt
streamlit
streamlit-authenticator
pyyaml
pdfplumber
docx2txt
bcrypt
